pyproject.toml
[tool.poetry]
name = "techtestcodingteam"
version = "0.1.0"
description = ""
authors = ["Takada <takada@nous.dev>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
django-environ = "^0.8.1"
psycopg2-binary = "^2.9.3"
django-model-utils = "^4.2.0"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.16.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
